backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
fpdf2==2.7.6
feedparser==6.0.10
requests==2.31.0
python-dotenv==1.0.0
alpha_vantage==2.3.1
# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
# Monitoring & Logging
structlog==23.2.0
# Security Headers
secure==0.3.0
